Add calculator.h
